ListBuilding: lannister units
Added the files I created before the repository was live. This is in a different structure than what was already published so we will sort out the similarities tomorrow.

Main change introduces the /api/v1/units endpoint which currently just implements lannister combat unit data. The builder format is mainly a timesaver, longterm I suspect we will be saving the data in resources as JSON.