More WIP
This is tonight's progress on list building & validation. The model seems to have coalesced better and I applied some suggestions from Femto.

A new validate endpoint is exposed to be able to test converting between a submitted list format, and using that to lookup the current unit balance data from JSON.