Most validators complete
I was able to get most of the validator functions defined. I moved the requires/excludes validators to be generated on a per-unit bases in order to create a closure around unit & combat unit (I still need to come to better names for these classes). Validators have been tied into the controller. I'm pretty happy with how the interface is coming together.